fix(cli): remove "No..." options for Foundation Model deployment
Bedrock is not actually deployed, but still requires Foundation Model deployment for
the inventory updates. This causes confusion with the "No..." based options in the CLI,
which are actually for "Cross-Account" control and advanced configuration, which can
just be handled by "Cross-Account" really.

By removing the "No..." options, this will remove confusion and always require updating
the inventory.