feat(corpus): use ManagedEmbeddings SageMakerEndpoint for embeddings
Remove the local embedding model in the corpus containers and replace embeddings
holistically with the SageMakerEndpoint managed embedding model that is configured
via the config.json.

While the model supports multiple embedding models, the application still
only support 1, which is denoted by the `default` property or the first model in the list.
The default is propagated through the system, and conceptually the
replacement for the local embedding model previously defined in the container.

BREAKING CHANGE: local embedding model within the corpus containers
has been removed. To change the embedding model, you need to configure
the `rag.embeddingModels` property in the `config.json`