feat(sdk): support browser logger variant with esm
Didn't end up using this for website usage of sdk, but that is where
it started and worked just for console. In the end browserified the other
needed node natives in website via webpack for full support which
negated the need for this, but still useful as workaround without webpack
in future.